tests/justhtml-tests/template_recovery.dat

#data
<template><table><colgroup></colgroup><script></script></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <script>
|   <body>

#data
<template><table><colgroup></colgroup><template></template></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <template>
|             content
|   <body>

#data
<template><table><colgroup><script></script></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <script>
|   <body>

#data
<template><table><colgroup><style></style></table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <table>
|           <colgroup>
|           <style>
|   <body>

#data
<template><tbody><colgroup></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|         <colgroup>
|   <body>

#data
<template><tbody></table><script></script></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|         <script>
|   <body>

#data
<template><tr></table><template></template></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|         <template>
|           content
|   <body>

#data
<template><tbody><table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|   <body>

#data
<template><tr><table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tr>
|   <body>

#data
<template><caption>x</caption><table></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <caption>
|           "x"
|   <body>

#data
<template><title>x</title><tbody></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <title>
|           "x"
|   <body>

#data
<template><title>x</title><tr></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <title>
|           "x"
|   <body>

#data
<template><title>x</title><td>x</template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <title>
|           "x"
|         "x"
|   <body>

#data
<template><optgroup></p></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <optgroup>
|           <p>
|   <body>

#data
<template></p><select></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <select>
|   <body>

#data
<template><tbody><table><tr></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <tbody>
|           <tr>
|   <body>

#data
<template><select><table></select><option></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <select>
|           <option>
|           <table>
|   <body>

#data
<template><base><tbody></template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <base>
|   <body>

#data
<template><noframes></noframes><td>x</template>
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <noframes>
|         "x"
|   <body>

#data
<template></template>&#32;
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|     " "
|   <body>

#data
<body><template></template>
#errors
(1,1): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|   <body>
|     <template>
|       content

#data
<template></br>
#errors
(1,10): expected-doctype-but-got-start-tag
(1,11): unexpected-end-tag

#document
| <html>
|   <head>
|     <template>
|       content
|   <body>

#data
<template><noscript></noscript>Y
#errors
(1,10): expected-doctype-but-got-start-tag

#document
| <html>
|   <head>
|     <template>
|       content
|         <noscript>
|         "Y"
|   <body>
